// ui/src/generated/types/updateUserGroupsQuotasParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */

export type UpdateUserGroupsQuotasParams = {
  /**
   * Number of groups
   */
  number?: number
  /**
   * Total duration of groups (milliseconds)
   */
  duration?: number
  /**
   * Number of repetitions per Group
   */
  repetitions?: number
}