// ui/src/generated/types/userGroupsQuotas.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * DeviceHub
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 1.3.10
 */
import type { UserGroupsQuotasAllocated } from './userGroupsQuotasAllocated'
import type { UserGroupsQuotasConsumed } from './userGroupsQuotasConsumed'

export type UserGroupsQuotas = {
  allocated?: UserGroupsQuotasAllocated
  consumed?: UserGroupsQuotasConsumed
  defaultGroupsDuration?: number
  defaultGroupsNumber?: number
  defaultGroupsRepetitions?: number
  repetitions?: number
}